alpaca_ids/requirements.txt
mysql-connector-python
pandas
numpy
hvac